rename: 모달 내용 컴포넌트 하나의 폴더로 이동
