+++ src/components/modal/modalContent/BasicModalContent/BasicModalContent.jsx
@@ -1,12 +1,14 @@
 import React, { useState } from 'react';
-import LogoutAlert from '../alert/LogoutAlert/LogoutAlert';
+import LogoutAlert from '../../alert/LogoutAlert/LogoutAlert';
 export default function BasicModalContent() {
   //alert창 열기
   const [isOpenAlert, setIsOpenAlert] = useState(false);
   const handleOpenAlert = () => {
-    setIsOpenAlert(!isOpenAlert);
+    setIsOpenAlert(true);
   };
-
+  const handleCloseAlert = () => {
+    setIsOpenAlert(false);
+  };
   return (
     <>
       <li>
@@ -15,7 +17,7 @@ export default function BasicModalContent() {
       <li>
         <button onClick={handleOpenAlert}>로그아웃</button>
       </li>
-      {isOpenAlert && <LogoutAlert handleOpenAlert={handleOpenAlert} />}
+      {isOpenAlert && <LogoutAlert handleCloseAlert={handleCloseAlert} />}
     </>
   );
 }
